encoding/thrift: Add mechanisms to handle requests via thriftrw's 'nowire' implementaiton
thriftrw's 'nowire' implementation (known as 'streaming' in the thriftrw
context) is an alternative wire-compatible implementation of the Thrift
protocol that bypasses the intermediate 'thriftrw.Wire' representation and
'streams' the protocol directly to the related objects.

This provides the mechanisms to handle 'Call's made via the 'nowire'
implementation instead of the existing 'Wire'-intermediary format.  Doing so
involved creating a new type, `thriftNoWireHandler`, that implements both a
`transport.UnaryHandler` and a `transport.OnewayHandler`.

These call into a new interface, `NoWireHandler` (which will be implemented by
generated code, for each procedure), that understands how to unpack the request
using thriftrw's 'nowire' primitives (`stream.RequestReadeer`), execute the
request, and return objects that understand what's involved in sending back a
response, if it's a Unary call.